+++ RELEASE_NOTES.md
@@ -1,3 +1,4 @@
 v0.99.1
 
+- node: Support SVR2 enclave migration.
 - Swift, Node: Don't log an error about unexpected server timestamps on unauthenticated connections.